Finish validations
